reorganized

+++ include/malloc.h
@@ -1,6 +1,6 @@
 #ifndef __MALLOC_H
 #define __MALLOC_H
 
-void *kmalloc(int sz);
+void *kmalloc(long sz);
 void kfree(void *);
 #endif
